+++ BE/src/main/java/com/longkubi/qlns/model/dto/search/TimeKeepingSearchDto.java
@@ -2,7 +2,6 @@ package com.longkubi.qlns.model.dto.search;
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
-import com.longkubi.qlns.model.dto.EmployeeDto;
 import com.longkubi.qlns.model.dto.PaymentSalaryDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -15,31 +14,21 @@ import java.util.UUID;
 @NoArgsConstructor
 @Getter
 @Setter
-@JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class,property = "id")
-public class TimeKeepingSearchDto extends SearchDto{
-
+@JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class, property = "id")
+public class TimeKeepingSearchDto extends SearchDto {
     private String code;//mã chấm công
 
     //private EmployeeDto employee;//nhân viên
-
     private UUID employeeId;
-
-    private Integer numberWorkDay;//Số Ngày Công Đi Làm
+    private Byte numberWorkDay;//Số Ngày Công Đi Làm
 
     //    @Column(name = "_leave")
     //    private int leave;
-
-    private Integer numberDayOff;//Số Ngày Nghỉ Có Phép
-
-    private Integer numberDayUnexcusedLeave;//Số Ngày Nghỉ Không Phép
-
-    private Integer month;//tháng
-
-    private Integer year;//năm
-
-    private Integer numberOvertimeHours; //Giờ Làm Thêm
-
+    private Byte numberDayOff;//Số Ngày Nghỉ Có Phép
+    private Byte numberDayUnexcusedLeave;//Số Ngày Nghỉ Không Phép
+    private Byte month;//tháng
+    private Short year;//năm
+    private Short numberOvertimeHours; //Giờ Làm Thêm
     private PaymentSalaryDto paymentSalary;
-
-    private Integer status;
+    private Byte status;
 }
